requirements.txt
pyqlib==0.9.5
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.0
pyyaml==6.0.1
tqdm==4.66.4
einops==0.8.0
